Move AppProfileIcon component to matters dir
