+++ pages/index.vue
@@ -86,7 +86,7 @@ import AppHeader1 from '~/components/elements/AppHeader1'
 import AppHeader2 from '~/components/elements/AppHeader2'
 import AppLinkButton from '~/components/elements/AppLinkButton'
 import AppPayPayButton from '~/components/matters/AppPayPayButton'
-import AppProfileIcon from '~/components/AppProfileIcon'
+import AppProfileIcon from '~/components/matters/AppProfileIcon'
 import AppSkillButton from '~/components/matters/AppSkillButton'
 import AppText from '~/components/elements/AppText'
 import AppWishListButton from '~/components/matters/AppWishListButton'
